Adding feature GEOTIFF to the EnTK pipeline
